Refactor module imports in xodium.index.ts and xodium.utils.ts
