+++ src/assets/modules/xodium.index.ts
@@ -1,4 +1,3 @@
 // xodium.index.ts
 import "xodium/typewriter";
-import "xodium/utils/github";
 import "xodium/utils";
