+++ src/assets/modules/xodium.utils.ts
@@ -1,18 +1,20 @@
 // xodium.utils.ts
+import { GithubService } from "xodium/utils/github";
 
-const clickEvent = "click";
-const focusoutEvent = "focusout";
+const CLICK_EVENT: string = "click";
+const FOCUS_OUT_EVENT: string = "focusout";
 
 class Utils {
   constructor() {
-    [clickEvent, focusoutEvent].forEach((eType) =>
+    [CLICK_EVENT, FOCUS_OUT_EVENT].forEach((eType) =>
       document.addEventListener(eType, (e) => {
         this.handleToggle(e, "data-toggle", "hidden");
       })
     );
-    document.addEventListener(clickEvent, (e) => {
+    document.addEventListener(CLICK_EVENT, (e) => {
       this.handleScroll(e, "data-scroll", "smooth");
     });
+    this.populateTeamGrid();
   }
 
   handleToggle = (e: Event, attr: string, classtype: string) => {
@@ -20,12 +22,12 @@ class Utils {
     if (target) {
       e.preventDefault();
       const element = document.getElementById(target);
-      if (e.type === clickEvent) {
+      if (e.type === CLICK_EVENT) {
         const isOpen = !element?.classList.contains(classtype);
         element?.classList.toggle(classtype);
         this.toggleArrow(isOpen);
       } else if (
-        e.type === focusoutEvent &&
+        e.type === FOCUS_OUT_EVENT &&
         !element?.classList.contains(classtype)
       ) {
         element?.classList.add(classtype);
@@ -48,6 +50,45 @@ class Utils {
       arrow.textContent = isOpen ? "▼" : "▲";
     }
   };
+
+  async populateTeamGrid(): Promise<void> {
+    const members = await GithubService.StoreOrgMembers();
+    const grid = document.querySelector(".team-grid");
+    if (grid) {
+      if (Array.isArray(members)) {
+        members.forEach((member) => {
+          const card = document.createElement("li");
+          card.classList.add("mb-4");
+          card.innerHTML = `
+              <div class="flex items-center gap-x-6">
+                <a href="${member.html_url}" target="_blank">
+                  <img
+                    class="h-16 w-16 rounded-full"
+                    src="${member.avatar_url}"
+                    alt="${member.login} picture"
+                  />
+                </a>
+                <div>
+                  <h3
+                    class="text-base font-semibold leading-7 tracking-tight text-gray-900 dark:text-slate-100"
+                  >
+                    ${member.login}
+                  </h3>
+                  <p class="text-sm font-semibold leading-6 text-indigo-600">
+                    ${member.role}
+                  </p>
+                </div>
+              </div>
+            `;
+          grid.appendChild(card);
+        });
+      } else {
+        console.error("Expected members to be an array, but got:", members);
+      }
+    } else {
+      console.error('Element with class ".team-grid" not found');
+    }
+  }
 }
 
 new Utils();
